back_end/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
plaid-python
python-dotenv
pymysql
requests
sendgrid
autogluon.timeseries
pandas
matplotlib
numpy
autogluon
openpyxl
tzdata